Experimental WMI events interceptor

+++ X7 Engine/PCC/inc/ProcessData.h
@@ -19,8 +19,8 @@ class ProcessData {
   const std::string& getPath();
 
   // Methods
-  int8_t setAffinity(uint64_t& affinity_mask);
-  int8_t getAffinity(uint64_t& affinity_mask);
+  int16_t setAffinity(uint64_t& affinity_mask);
+  int16_t getAffinity(uint64_t& affinity_mask);
 
  private:
   HANDLE handle;
